+++ bungee/src/main/java/com/jannik_kuehn/loritimebungee/BungeePluginMessenger.java
@@ -6,9 +6,9 @@ import net.md_5.bungee.api.event.PluginMessageEvent;
 import net.md_5.bungee.api.plugin.Listener;
 import net.md_5.bungee.event.EventHandler;
 
-public class BungeePluginMessage extends PluginMessaging implements Listener {
+public class BungeePluginMessenger extends PluginMessaging implements Listener {
 
-    public BungeePluginMessage(final LoriTimeBungee bungeePlugin) {
+    public BungeePluginMessenger(final LoriTimeBungee bungeePlugin) {
         super(bungeePlugin.getPlugin());
     }