+++ bungee/src/main/java/com/jannik_kuehn/loritimebungee/LoriTimeBungee.java
@@ -77,7 +77,7 @@ public class LoriTimeBungee extends Plugin {
     private void enableRemainingFeatures() {
         if (loriTimePlugin.isAfkEnabled()) {
             getProxy().registerChannel("loritime:afk");
-            getProxy().getPluginManager().registerListener(this, new BungeePluginMessage(this));
+            getProxy().getPluginManager().registerListener(this, new BungeePluginMessenger(this));
             loriTimePlugin.enableAfkFeature(new MasteredAfkPlayerHandling(loriTimePlugin));
         }
     }